FIX: Fix incorrect data type for id from int to long
DB use bigint for id

+++ Models/Holiday.cs
@@ -2,18 +2,18 @@
 {
     public class Holiday
     {
-        public int? HolidayId { get; set; }
-        public int? UserId { get; set; }
+        public long? HolidayId { get; set; }
+        public long? UserId { get; set; }
         public string? Username { get; set; }
         public required DateOnly Date {  get; set; }
 
 
-        public static Holiday NewHoliday(int userId, DateOnly date)
+        public static Holiday NewHoliday(long userId, DateOnly date)
         {
             return new Holiday { UserId = userId, Date = date };
         }
 
-        public static Holiday NewRegisteredHoliday(int holidayId, int userId, string username, DateOnly date)
+        public static Holiday NewRegisteredHoliday(long holidayId, long userId, string username, DateOnly date)
         {
             return new Holiday { HolidayId = holidayId, UserId = userId, Username = username, Date = date };
         }
